Refactored attack related function out of MatrixFlow
