+++ src/module/tests/hooks/MatrixHooks.ts
@@ -1,10 +1,10 @@
 import { OverwatchFlow } from '@/module/flows/OverwatchFlow';
-import { MatrixFlow } from '../../flows/MatrixFlow';
 import { MatrixTestDataFlow } from '../flows/MatrixTestDataFlow';
 import { SuccessTest } from '../SuccessTest';
 import { SR5Actor } from '@/module/actor/SR5Actor';
 import { DamageType } from '@/module/types/item/Action';
 import { MatrixDamageFlow } from '@/module/flows/MatrixDamageFlow';
+import { MatrixAttackFlow } from '@/module/flows/MatrixAttackFlow';
 
 /**
  * Matrix related Success Test hook implementations.
@@ -44,11 +44,11 @@ export const MatrixHooks = {
     },
 
     /**
-     * After a test has been executed, determine if  the 1 damage that you get from failing ATTACK based actions SR 231
+     * After a test has been executed, determine if the 1 damage that you get from failing ATTACK based actions SR 231
      * @param test
      */
     onAfterTestComplete_HandleMatrixFailedAttack: async function(test: SuccessTest) {
-        await MatrixFlow.determineMatrixFailedAttack(test);
+        await MatrixAttackFlow.determineMatrixFailedAttack(test);
     },
 
     /**
